Thickness scan
Since we are writing CDR, we probably should answer this question to the next leading order precision, which involves eta-dependency, the tilting and adjustments of tower density (i.e. less fiber than 1-D SPACAL). Please note 14.4 cm is the reserved space for SPACAL blocks, and the block length various from 13.5 cm in the central rapidities to much longer ones in the forward directions. 

Of course at NLO, the answer also becomes more complex. Attached is the material scan with the most recent EMCal model. This is calculated with only SPACAL tower, light guides and side sector walls, i.e. without counting the strong back, tower support structure, cooling and electronics materials. While maintaining the minimal coverage of 18 X-0 in the eta directions, the average radiation length is 20.1 X0. The forward towers are significantly thicker to compensate for the stair-gap leakages and higher total shower energy. The nuclear interaction length is larger than 0.75 Lambda0 in the polar direction with an average of 0.8 Lambda0. In the azimuthal direction, there is obvious dip in the sector boundary, as expected from the sector assembly clearance. Due to our tilting design, nothing touches zero (except exactly eta = 0, z = 0).