Updates to KFParticle package
Implemented a new class which contains particle information. Currently has a map to get the mass of various particles used in the reconstruction such as pions, kaons, W's, B's

Added a new option to reconstruct the charge conjugate state or not. Works for D0 -> Kpi

Changes the output branch names to use the mother or intermediate names instead of "mother" and "intermediate_[n]". Easier to tell what variables relate to J/psi of Bs0 for example.