Fixed truth matching in KFParticle
This update corrects a bug in the truth matching that would crash the program. If the user sets the truth matching flag to true (kfparticle->doTruthMatching(1);) then track and associated decay vertex information will be written to the output. I can't seem to get the parent truth information just yet as the parent IDs are set to zero. Iterating over all truth particles in the event and printing their particle IDs doesnt reveal the mothers either (tested with D0->Kpi events, D0 has a true ID of +/- 421 which never appeared in the particle map).