First iteration of heavy flavor trigger
This PR adds a VERY basic heavy flavor trigger. This is part of a study on a hardware trigger. To this end, the triggers here use simple quantities: track-vertex DCA, track-track DCA, track pT and nHits in INTT.

There are four triggers that can be enabled: a one track trigger, a two track trigger (which calls the one track trigger), a low multiplicity trigger and a high multiplicity trigger. You can enable multiple triggers but they are all required to trigger in your event to enable processing so be careful for now. We also trigger independent of your signal (TIS) not on the signal (TOS). There is also no appending back to the node tree here, if your trigger on the event you process the event, if not then you skip it. This means you have to run the trigger everytime you run Fun4All (no pre-processing available).

This trigger works on reconstructed tracks, I would like to keep this as an option and work on triggering on raw hits. THis means we can port the trigger into full reconstruction if needed later.

Example (from Fun4All_KFParticle_advanced.C):

R__LOAD_LIBRARY(libhftrigger.so)

  bool use_trigger = true;
  if (use_trigger)
  {
    HFTrigger* myTrigger = new HFTrigger("myTestTrigger");
    myTrigger->Verbosity(verbosity);
    myTrigger->requireOneTrackTrigger(true);
    myTrigger->requireTwoTrackTrigger(true);
    myTrigger->requireLowMultiplicityTrigger(true);
    se->registerSubsystem(myTrigger);
  }