KFParticle: Changes to advanced example
I made a few changes to the Fun4All_KFParticle_advanced.C macro. The code to run acts tracking now calls an ACTS macro (in its own folder NOT from the macros repository, change the include to use the macros version). The reason for using a local copy is that it is taken from the MDC production and the macros version hasn't been updated to match yet (I think).

Thanks to the work from Joe, I added an option to rerun the ACTS final vertexing from the SvtxTrackMap. We discovered in our MDC production that pythia8 makes lots of truth vertex copies which got passed into our final vertex map. This has a minimal impact as the truth vertices are all on top of each other but some of the calcualtions and uncertainties could be incorrect. See https://indico.bnl.gov/event/10855/contributions/45863/attachments/32826/52483/MDC1_C_Dean_20210222.pdf