KFParticle: Changed examples to find files over dCache.
With files being moved around, the condor submission scripts using a hard path to find the DST is a bad idea. Also, with close to 40k charm DSTs, running a single job per file is also wasteful. I've changed the python scripts to make a list of DSTs from either sample (the number of files in each list is user configurable) and then modified Fun4All_KFParticle_singleFile.C (renamed to Fun4All_KFParticle_condorJob.C) to read this list and find the relevant DSTs through dCache.

I also added Joe's fix for the PVs to this PR and defaulted it to run in Fun4All_KFParticle_condorJob.C