Updates to heavy flavor tools
Finalised DecayFinder to look for specific decays in the HepMC record. The module skips over certain resonances (unless the user has them in their decay descriptor) and can skip over photons and pi0. If the user wishes, they can trigger on these decays to purify their samples.

Added an anti-trigger. The user passes the module a list of particles and if they exist in that events HepMC record, the rest of the event processing if terminated. Useful for background studies where you can delete, for example, Lambda_c events

Updated KFParticle example macros to take advatnge of the new packages and the new decay parser. Also added a list of min-bias files and an option in makeCOndorJobs.py for min-bias production